# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Api
 * M4Chat API specification
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterRequest {
  /**
     * @minLength 3
     * @maxLength 100
     */
  username: string;
  /** @minLength 6 */
  password: string;
  email?: string;
  phone?: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface User {
  id: number;
  username: string;
  email?: string;
  phone?: string;
  avatarUrl?: string;
  bio?: string;
  isOnline?: boolean;
  lastSeenAt?: string;
  twoFAEnabled?: boolean;
}

export interface AuthResponse {
  user: User;
  token: string;
  refreshToken?: string;
}

export interface UserPublic {
  id: number;
  username: string;
  avatarUrl?: string;
  bio?: string;
  isOnline?: boolean;
  lastSeenAt?: string;
}

export interface UpdateProfileRequest {
  email?: string;
  phone?: string;
  bio?: string;
  avatarUrl?: string;
}

export interface Session {
  id: number;
  deviceInfo?: string;
  ipAddress?: string;
  createdAt?: string;
  expiresAt?: string;
}

export type ChatType = typeof ChatType[keyof typeof ChatType];


export const ChatType = {
  private: 'private',
  group: 'group',
  channel: 'channel',
} as const;

export interface Chat {
  id: number;
  name?: string;
  type: ChatType;
  photo?: string;
  inviteLink?: string;
  pinnedMessageId?: number;
  isArchived?: boolean;
  createdBy?: number;
  createdAt?: string;
}

export interface MessagePreview {
  id?: number;
  content?: string;
  messageType?: string;
  senderId?: number;
  senderName?: string;
  createdAt?: string;
}

export type ChatItem = Chat & {
  role?: string;
  unreadCount?: number;
  lastMessage?: MessagePreview;
};

export type ChatMemberRole = typeof ChatMemberRole[keyof typeof ChatMemberRole];


export const ChatMemberRole = {
  creator: 'creator',
  admin: 'admin',
  member: 'member',
  restricted: 'restricted',
} as const;

export interface ChatMember {
  id: number;
  userId: number;
  username?: string;
  avatarUrl?: string;
  role: ChatMemberRole;
  joinedAt?: string;
}

export type ChatDetail = Chat & {
  members?: ChatMember[];
};

export type CreateChatRequestType = typeof CreateChatRequestType[keyof typeof CreateChatRequestType];


export const CreateChatRequestType = {
  private: 'private',
  group: 'group',
  channel: 'channel',
} as const;

export interface CreateChatRequest {
  name?: string;
  type: CreateChatRequestType;
  participantIds?: number[];
}

export type MessageMessageType = typeof MessageMessageType[keyof typeof MessageMessageType];


export const MessageMessageType = {
  text: 'text',
  image: 'image',
  file: 'file',
  voice: 'voice',
  video: 'video',
  poll: 'poll',
} as const;

export interface Reaction {
  emoji?: string;
  count?: number;
  mine?: boolean;
}

export interface Message {
  id: number;
  chatId: number;
  senderId?: number;
  senderName?: string;
  senderAvatar?: string;
  content?: string;
  messageType: MessageMessageType;
  mediaUrl?: string;
  replyTo?: number;
  isEdited?: boolean;
  isDeleted?: boolean;
  createdAt: string;
  reactions?: Reaction[];
}

export interface SendMessageRequest {
  content: string;
  messageType?: string;
  mediaUrl?: string;
  replyTo?: number;
}

export interface FileItem {
  id: number;
  fileKey?: string;
  fileName?: string;
  mimeType?: string;
  size?: number;
  url: string;
  createdAt?: string;
}

export type RefreshTokenBody = {
  refreshToken: string;
};

export type Setup2FA200 = {
  secret?: string;
  qrCodeUrl?: string;
};

export type Verify2FABody = {
  code: string;
};

export type Verify2FA200 = {
  verified?: boolean;
};

export type SearchUsersParams = {
/**
 * @minLength 2
 */
q: string;
};

export type AddChatMemberBody = {
  userId: number;
};

export type GenerateInviteLink200 = {
  inviteLink?: string;
};

export type GetMessagesParams = {
limit?: number;
offset?: number;
};

export type EditMessageBody = {
  content: string;
};

export type AddReactionBody = {
  emoji: string;
};

export type PinMessageBody = {
  pinned?: boolean;
};

export type SearchMessagesParams = {
q: string;
chatId?: number;
};

export type UploadFileBody = {
  file: Blob;
  chatId?: number;
};