# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: M4Chat API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
        "409":
          description: Username already taken
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: New tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/devices:
    get:
      operationId: getDevices
      tags: [auth]
      summary: List active sessions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Sessions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"
  /auth/devices/{id}/revoke:
    delete:
      operationId: revokeDevice
      tags: [auth]
      summary: Revoke a session
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Revoked
  /auth/2fa/setup:
    post:
      operationId: setup2FA
      tags: [auth]
      summary: Setup 2FA TOTP
      security:
        - bearerAuth: []
      responses:
        "200":
          description: QR code and secret
          content:
            application/json:
              schema:
                type: object
                properties:
                  secret:
                    type: string
                  qrCodeUrl:
                    type: string
  /auth/2fa/verify:
    post:
      operationId: verify2FA
      tags: [auth]
      summary: Verify 2FA code
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                type: object
                properties:
                  verified:
                    type: boolean

  # Users
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 100
        password:
          type: string
          minLength: 6
        email:
          type: string
          format: email
        phone:
          type: string
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
        refreshToken:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        phone:
          type: string
        avatarUrl:
          type: string
        bio:
          type: string
        isOnline:
          type: boolean
        lastSeenAt:
          type: string
          format: date-time
        twoFAEnabled:
          type: boolean
      required:
        - id
        - username
...
    Session:
      type: object
      properties:
        id:
          type: integer
        deviceInfo:
          type: string
        ipAddress:
          type: string
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
      required:
        - id
...