# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Api
 * M4Chat API specification
 * OpenAPI spec version: 0.2.0
 */

export interface User {
  id: number;
  username: string;
  email?: string;
  phone?: string;
  avatarUrl?: string;
  bio?: string;
  isOnline?: boolean;
  lastSeenAt?: Date;
  twoFAEnabled?: boolean;
}